Move exception classes to 'exceptions' package.
