+++ src/main/java/com/github/dakusui/actionunit/ActionUnit.java
@@ -1,5 +1,6 @@
 package com.github.dakusui.actionunit;
 
+import com.github.dakusui.actionunit.exceptions.ActionException;
 import com.google.common.base.Function;
 import com.google.common.collect.Lists;
 import org.junit.Test;
