+++ src/main/java/com/github/dakusui/actionunit/exceptions/Abort.java
@@ -1,4 +1,4 @@
-package com.github.dakusui.actionunit;
+package com.github.dakusui.actionunit.exceptions;
 
 import com.github.dakusui.actionunit.actions.Retry;
