+++ src/test/java/com/github/dakusui/actionunit/ActionExceptionTest.java
@@ -1,5 +1,6 @@
 package com.github.dakusui.actionunit;
 
+import com.github.dakusui.actionunit.exceptions.ActionException;
 import org.junit.Test;
 
 import java.io.IOException;
